+++ CHANGES.rst
@@ -10,6 +10,11 @@ Bug fixes
 - Fix crash when pytest-xdist is installed but disabled.
   (Thanks to `@mgorny <https://github.com/mgorny>`_ for the PR.)
 
+Features
+++++++++
+
+- Added option `--rerun-except` to rerun failed tests those are other than the mentioned Error.
+
 
 10.2 (2021-09-17)
 -----------------